refactor key apikey in intellij
